requirements.txt
absl-py==2.1.0
accelerate==0.31.0
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
blosc2==2.0.0
braceexpand==0.1.7
cached-property==1.5.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
chex==0.1.7
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.16.0
clu==0.0.12
cmake==3.27.6
colorama==0.4.6
comm==0.2.2
confection==0.1.4
contextlib2==21.6.0
contexttimer==0.3.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
datasets==2.20.0
debugpy==1.8.1
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
diffusers==0.16.0
dill==0.3.8
distlib==0.3.8
dm-tree==0.1.8
easydict==1.9
einops==0.7.0
einops-exts==0.0.4
entrypoints==0.4
etils==1.3.0
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.4
fastapi==0.112.0
fastjsonschema==2.19.1
ffmpy==0.4.0
filelock==3.13.3
flash-attn==2.5.9.post1
flatbuffers==24.3.25
flax==0.7.2
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
gast==0.4.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core
google-api-python-client==2.151.0
google-auth==2.23.4
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos
gradio
gradio_client==1.3.0
gspread==6.1.4
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
identify==2.5.35
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
imagenetv2_pytorch @ git+https://github.com/modestyachts/ImageNetV2_pytorch@14d4456c39fe7f02a665544dd9fc37c1a5f8b635
importlib_metadata==8.4.0
importlib_resources==6.4.0
inflection==0.5.1
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.29.4
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
itables==2.1.4
jax==0.4.13
jaxlib==0.4.13
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_packaging==0.12.3
jupyter_server==2.13.0
jupyter_server_fileid==0.9.1
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kaggle==1.6.8
keras==2.13.1
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
libclang==18.1.1
lightning-utilities==0.11.2
lit==17.0.2
lmdb==1.5.1
loguru==0.7.2
loralib==0.1.2
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
ml_collections==0.1.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multimodal==0.0.13
multiprocess==0.70.16
murmurhash==1.0.10
natsort==8.4.0
nbclassic==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.6
notebook_shim==0.2.4
numexpr==2.8.6
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
nvidia-tensorrt==99.0.0
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
open-flamingo==2.0.1
open3d
open_clip_torch==2.26.1
opencv-python==4.8.0.74
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
openprompt
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.2.3
orjson==3.10.7
overrides==7.7.0
packaging==23.2
packg==0.16.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.12.1
peft==0.10.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plotly==5.20.0
portalocker==2.8.2
pre-commit==3.5.0
preshed==3.0.9
prettytable==3.10.0
proglog==0.1.10
promcse==0.0.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus
protobuf
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyinstrument==4.7.2
pyparsing==3.1.2
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==24.0.1
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.34.0
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rouge==1.0.0
rpds-py==0.18.0
rsa==4.9
ruff==0.5.7
sacrebleu==2.4.1
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.0.1
sentencepiece==0.1.98
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
spacy
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
streamlit
sympy==1.12
tables==3.8.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard
tensorboard-data-server==0.7.2
tensorboardX
tensorflow
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
tensorrt==10.0.1
tensorrt-cu12==10.0.1
tensorrt-cu12-bindings==10.0.1
tensorrt-cu12-libs==10.0.1
tensorstore==0.1.45
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==3.4.0
tifffile==2023.7.10
timm==1.0.8
tinycss2==1.2.1
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1
torch-tb-profiler==0.4.3
torchaudio
torchmetrics==1.4.0.post0
torchvision
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
transformers==4.46.3
triton
typedparser==0.16.2
typer
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.5
virtualenv==20.25.1
visiontext==0.16.4
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.13
weasel
webcolors==1.13
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xxhash==3.4.1
y-py==0.6.2
yacs==0.1.8
yarl==1.9.4
ypy-websocket==0.8.4
zipp==3.18.1
zstandard==0.23.0
